+++ src/main/java/com/github/games647/lagmonitor/commands/StackTraceCommand.java
@@ -1,18 +1,22 @@
 package com.github.games647.lagmonitor.commands;
 
 import com.github.games647.lagmonitor.LagMonitor;
+import com.google.common.collect.Lists;
+import java.lang.management.ManagementFactory;
+import java.lang.management.ThreadInfo;
+import java.util.Collections;
 
+import java.util.List;
 import java.util.Map;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
+import org.bukkit.command.TabExecutor;
 
-public class StackTraceCommand implements CommandExecutor {
+public class StackTraceCommand implements TabExecutor {
 
     private static final ChatColor PRIMARY_COLOR = ChatColor.DARK_AQUA;
-    private static final ChatColor SECONDARY_COLOR = ChatColor.GRAY;
 
     private final LagMonitor plugin;
 
@@ -61,5 +65,20 @@ public class StackTraceCommand implements CommandExecutor {
 
             sender.sendMessage(PRIMARY_COLOR + className + '.' + methodName + ':' + ChatColor.DARK_PURPLE + line);
         }
+    }
+
+    @Override
+    public List<String> onTabComplete(CommandSender sender, Command command, String alias, String[] args) {
+        List<String> result = Lists.newArrayList();
+        String lastArg = args[args.length - 1];
+        ThreadInfo[] threads = ManagementFactory.getThreadMXBean().dumpAllThreads(false, false);
+        for (ThreadInfo thread : threads) {
+            if (thread.getThreadName().startsWith(lastArg)) {
+                result.add(thread.getThreadName());
+            }
+        }
+
+        Collections.sort(result, String.CASE_INSENSITIVE_ORDER);
+        return result;
     }
 }